requirements.txt
# Core
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.16

# Validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# HTTP Client
httpx==0.28.1

# Redis
redis==5.0.1
aioredis==2.0.1

# AI/LLM dependencies
crewai==0.30.11
langgraph==0.0.69
langchain==0.2.3
langchain-openai==0.1.7
openai==1.30.5

# Development
pytest==8.3.4
pytest-asyncio==0.25.0
black==24.10.0
flake8==7.1.1